Using the new Config system. Restructured to fix cycles
